requirements.txt
python-dotenv
qiskit
qiskit-aer
qiskit-ibm-runtime
networkx
matplotlib
numpy
ipykernel
pennylane
dwave-ocean-sdk